Update gen_packets.c fix scale

+++ gen_packets.c
@@ -902,7 +902,7 @@ int audio_put (int a, int c)
 	}
 	else {
 	  byte_count++;
-	  c *= (amplitude/100f);
+	  c *= ((float)amplitude/100.0);
 	  if (c > 32767) c = 32767;
 	  if (c < -32767) c = -32767;		
 	  return (putc(c, out_fp));
